lots of fixes

+++ scripts/helpers.py
@@ -45,11 +45,58 @@ def format_scala(files):
         subprocess.call(["java", "-jar", os.path.join(BASE_PATH, "../tools/scalafmt.jar"), "-i", "-s", "defaultWithAlign", "--alignByArrowEnumeratorGenerator", "true", "--continuationIndentDefnSite","2", "--continuationIndentCallSite", "2", "--bestEffortInDeeplyNestedCode", "--maxColumn", "80", "-f", d], stdout=subprocess.PIPE)
         #subprocess.call(["java", "-jar", os.path.join(BASE_PATH, "../tools/scalafmt.jar"), "-i", "-s", "default", "--maxColumn", "80", "-f", d], stdout=subprocess.PIPE)
 
-def write(name, content):
+def confirm(prompt=None, resp=False):
+    """prompts for yes or no response from the user. Returns True for yes and
+    False for no.
+
+    'resp' should be set to the default value assumed by the caller when
+    user simply types ENTER.
+
+    >>> confirm(prompt='Create Directory?', resp=True)
+    Create Directory? [y]|n:
+    True
+    >>> confirm(prompt='Create Directory?', resp=False)
+    Create Directory? [n]|y:
+    False
+    >>> confirm(prompt='Create Directory?', resp=False)
+    Create Directory? [n]|y: y
+    True
+
+    """
+
+    if prompt is None:
+        prompt = 'Confirm'
+
+    if resp:
+        prompt = '%s [%s]|%s: ' % (prompt, 'y', 'n')
+    else:
+        prompt = '%s [%s]|%s: ' % (prompt, 'n', 'y')
+
+    while True:
+        ans = raw_input(prompt)
+        if not ans:
+            return resp
+        if ans not in ['y', 'Y', 'n', 'N']:
+            print 'please enter y or n.'
+            continue
+        if ans == 'y' or ans == 'Y':
+            return True
+        if ans == 'n' or ans == 'N':
+            return False
+
+def write(name, content, prompt_overwrite=False):
     #verify that the base path exists
     base = os.path.dirname(name)
     if not os.path.exists(base):
         os.makedirs(base)
-    fd = open(name, "w+")
-    fd.write(content)
-    fd.close()
+
+    if os.path.exists(name) and prompt_overwrite:
+        #prompt for confirmation
+        if confirm("Overwrite %s?" % (name)):
+            fd = open(name, "w+")
+            fd.write(content)
+            fd.close()
+    else:
+        fd = open(name, "w+")
+        fd.write(content)
+        fd.close()
